// src/types/prismaTypes.d.ts
export type Property = $Result.DefaultSelection<Prisma.$PropertyPayload>;
// ...
export namespace $Enums {
  export const Highlight: {
    HighSpeedInternetAccess: "HighSpeedInternetAccess";
    WasherDryer: "WasherDryer";
    AirConditioning: "AirConditioning";
    Heating: "Heating";
    SmokeFree: "SmokeFree";
    CableReady: "CableReady";
    SatelliteTV: "SatelliteTV";
    DoubleVanities: "DoubleVanities";
    TubShower: "TubShower";
    Intercom: "Intercom";
    SprinklerSystem: "SprinklerSystem";
    RecentlyRenovated: "RecentlyRenovated";
    CloseToTransit: "CloseToTransit";
    GreatView: "GreatView";
    QuietNeighborhood: "QuietNeighborhood";
  };

  export type Highlight = (typeof Highlight)[keyof typeof Highlight];

  export const Amenity: {
    WasherDryer: "WasherDryer";
    AirConditioning: "AirConditioning";
    Dishwasher: "Dishwasher";
    HighSpeedInternet: "HighSpeedInternet";
    HardwoodFloors: "HardwoodFloors";
    WalkInClosets: "WalkInClosets";
    Microwave: "Microwave";
    Refrigerator: "Refrigerator";
    Pool: "Pool";
    Gym: "Gym";
    Parking: "Parking";
    PetsAllowed: "PetsAllowed";
    WiFi: "WiFi";
  };

  export type Amenity = (typeof Amenity)[keyof typeof Amenity];

  export const PropertyType: {
    Rooms: "Rooms";
    Tinyhouse: "Tinyhouse";
    Apartment: "Apartment";
    Villa: "Villa";
    Townhouse: "Townhouse";
    Cottage: "Cottage";
  };

  export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType];

  export const ApplicationStatus: {
    Pending: "Pending";
    Denied: "Denied";
    Approved: "Approved";
  };

  export type ApplicationStatus =
    (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

  export const PaymentStatus: {
    Pending: "Pending";
    Paid: "Paid";
    PartiallyPaid: "PartiallyPaid";
    Overdue: "Overdue";
  };

  export type PaymentStatus =
    (typeof PaymentStatus)[keyof typeof PaymentStatus];
}
// ...
export namespace Prisma {
// ...
  export const type: unique symbol;
// ...
  export type $PropertyPayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Property";
    objects: {
      location: Prisma.$LocationPayload<ExtArgs>;
      manager: Prisma.$ManagerPayload<ExtArgs>;
      leases: Prisma.$LeasePayload<ExtArgs>[];
      applications: Prisma.$ApplicationPayload<ExtArgs>[];
      favoritedBy: Prisma.$TenantPayload<ExtArgs>[];
      tenants: Prisma.$TenantPayload<ExtArgs>[];
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        name: string;
        description: string;
        pricePerMonth: number;
        securityDeposit: number;
        applicationFee: number;
        photoUrls: string[];
        amenities: $Enums.Amenity[];
        highlights: $Enums.Highlight[];
        isPetsAllowed: boolean;
        isParkingIncluded: boolean;
        beds: number;
        baths: number;
        squareFeet: number;
        propertyType: $Enums.PropertyType;
        postedDate: Date;
        averageRating: number | null;
        numberOfReviews: number | null;
        locationId: number;
        managerCognitoId: string;
      },
      ExtArgs["result"]["property"]
    >;
    composites: {};
  };
// ...
  export type $ManagerPayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Manager";
    objects: {
      managedProperties: Prisma.$PropertyPayload<ExtArgs>[];
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        cognitoId: string;
        name: string;
        email: string;
        phoneNumber: string;
      },
      ExtArgs["result"]["manager"]
    >;
    composites: {};
  };
// ...
  export type $TenantPayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Tenant";
    objects: {
      properties: Prisma.$PropertyPayload<ExtArgs>[];
      favorites: Prisma.$PropertyPayload<ExtArgs>[];
      applications: Prisma.$ApplicationPayload<ExtArgs>[];
      leases: Prisma.$LeasePayload<ExtArgs>[];
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        cognitoId: string;
        name: string;
        email: string;
        phoneNumber: string;
      },
      ExtArgs["result"]["tenant"]
    >;
    composites: {};
  };
// ...
  export type $LocationPayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Location";
    objects: {
      properties: Prisma.$PropertyPayload<ExtArgs>[];
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        address: string;
        city: string;
        state: string;
        country: string;
        postalCode: string;
      },
      ExtArgs["result"]["location"]
    >;
    composites: {};
  };
// ...
  export type $ApplicationPayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Application";
    objects: {
      property: Prisma.$PropertyPayload<ExtArgs>;
      tenant: Prisma.$TenantPayload<ExtArgs>;
      lease: Prisma.$LeasePayload<ExtArgs> | null;
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        applicationDate: Date;
        status: $Enums.ApplicationStatus;
        propertyId: number;
        tenantCognitoId: string;
        name: string;
        email: string;
        phoneNumber: string;
        message: string | null;
        leaseId: number | null;
      },
      ExtArgs["result"]["application"]
    >;
    composites: {};
  };
// ...
  export type $LeasePayload<
    ExtArgs extends $Extensions.InternalArgs = $Extensions.DefaultArgs
  > = {
    name: "Lease";
    objects: {
      property: Prisma.$PropertyPayload<ExtArgs>;
      tenant: Prisma.$TenantPayload<ExtArgs>;
      application: Prisma.$ApplicationPayload<ExtArgs> | null;
      payments: Prisma.$PaymentPayload<ExtArgs>[];
    };
    scalars: $Extensions.GetPayloadResult<
      {
        id: number;
        startDate: Date;
        endDate: Date;
        rent: number;
        deposit: number;
        propertyId: number;
        tenantCognitoId: string;
      },
      ExtArgs["result"]["lease"]
    >;
    composites: {};
  };
// ...
}